System runs by passing a configuration file to run_forecast_system.
Many changes to do this including:
Pass dictionaries into forecast_system, as a result variables had to be
renamed.
Reconstruct param_dic so that it can be saved to netcdf file.
Create results file path in run_forecast_system.
Create a configure yaml file and read it in run_forecast_system.
Add argparse to run_forecast_system so that it can be run from command
line.