Error stats can be returned for multiple days.
New functions were written to analyze forecasts on different
days. The biggest change from the single day case is the need to
combine netcdf files corresponding to different days. This means that
they are in different folders and are a different size. Since the
files are in different folders a list must be passed instead of a path
with a wildcard. Since the domain sizes are different all the files
must be cropped to the error domain in a prepossessing step.